Make application package, split applications
